pannuke_project GITHUB/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.10
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
Pillow>=10.0.0

# Optional but commonly used
opencv-python>=4.8.0
matplotlib>=3.8.0
